Remove duplication in localization tests
